requirements.txt
Flask==2.3.3
Flask-Login==0.7.0
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.6
python-dotenv
gunicorn